Component - MeshRenderer 클래스 생성

+++ GameCoding/IndexBuffer.cpp
@@ -2,14 +2,12 @@
 #include "IndexBuffer.h"
 
 IndexBuffer::IndexBuffer(ComPtr<ID3D11Device> device)
-	: _device(device)
+	:_device(device)
 {
-
 }
 
 IndexBuffer::~IndexBuffer()
 {
-
 }
 
 void IndexBuffer::Create(const vector<uint32>& indices)
@@ -21,7 +19,7 @@ void IndexBuffer::Create(const vector<uint32>& indices)
 	ZeroMemory(&desc, sizeof(desc));
 	desc.Usage = D3D11_USAGE_IMMUTABLE;
 	desc.BindFlags = D3D11_BIND_INDEX_BUFFER;
-	desc.ByteWidth = (uint32)(_stride * _count);
+	desc.ByteWidth = _stride * _count;
 
 	D3D11_SUBRESOURCE_DATA data;
 	ZeroMemory(&data, sizeof(data));
